Camera fixes and Zooms

+++ src/at/htblakaindorf/AHIF18/Entity/Entity.java
@@ -4,7 +4,6 @@ import java.awt.image.BufferedImage;
 
 public class Entity {
 
-    public int worldX, worldY;
-    public int speed;
-
+    public double worldX, worldY;
+    public double speed;
 }
